Solution to Calculate Total and Average Marks of Students

+++ Calculate Total and Average Marks of Students@@67877d61dc107486ad5c2746/code.c
@@ -2,7 +2,7 @@
 int main(){
     int a;
     scanf("%d",&a);
-    int total_marks=0;
+    float total_marks=0;
     for(int i=1;i<=a;i++){
         int b;
         char c[20];
